+++ __init__.py
@@ -1,4 +1,6 @@
-from .toolkitbrazil import rngCNPJ, rngCNPJfiliais, rngCPF, valCNPJ, valCPF, valCPFuf, ufCapital, ufCPF, ufDDD, strClean
+from .toolkitbrazil import rngCNPJ, rngCNPJfiliais, rngCPF, valCNPJ, valCPF, valCPFuf, ufCapital, ufCPF, ufDDD, strClean, valBissexto, dtPascoa, valPascoa, dtCarnaval, valCarnaval, rngPlaca, valEmail, ufCep
 
 __all__ = ['rngCNPJ', 'rngCNPJfiliais', 'rngCPF', 'valCNPJ',
-           'valCPF', 'valCPFuf', 'ufCapital', 'ufCPF', 'ufDDD', 'strClean']
+           'valCPF', 'valCPFuf', 'ufCapital', 'ufCPF', 'ufDDD', 'strClean',
+           'valBissexto', 'dtPascoa', 'valPascoa', 'dtCarnaval', 'valCarnaval',
+           'rngPlaca', 'valEmail', 'ufCep']